backend/requirements.txt
python-dotenv
PyPDF2
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytesseract
pillow
opencv-python
python-multipart
